refactor: 화면별 파일 분리 #6
